+++ src/Products/PloneMeeting/MeetingItem.py
@@ -4858,9 +4858,22 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
 
     def getCustomAdviceMessageFor(self, advice):
         '''See doc in interfaces.py.'''
+        customAdviceMessage = None
+        if advice['hidden_during_redaction']:
+            context = self.getSelf()
+            if advice['advice_editable']:
+                customAdviceMessage = translate(
+                    'hidden_during_redaction',
+                    domain='PloneMeeting',
+                    context=context.REQUEST)
+            else:
+                customAdviceMessage = translate(
+                    'considered_not_given_hidden_during_redaction',
+                    domain='PloneMeeting',
+                    context=context.REQUEST)
         return {'displayDefaultComplementaryMessage': True,
                 'displayAdviceReviewState': False,
-                'customAdviceMessage': None}
+                'customAdviceMessage': customAdviceMessage}
 
     def _getInsertOrder(self, cfg):
         '''When inserting an item into a meeting, several "methods" are
@@ -5376,9 +5389,8 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
         res = []
         for adviserUid in adviserUids:
             # adviserId could not exist if we removed an inherited initiative advice for example
-            if not predecessor.adviceIndex.get(adviserUid, None):
-                continue
-            if (optional and not predecessor.adviceIndex[adviserUid]['optional']):
+            if not predecessor.adviceIndex.get(adviserUid, None) or \
+               not predecessor.adviceIndex[adviserUid]['optional'] == optional:
                 continue
             res.append(
                 {'org_uid': predecessor.adviceIndex[adviserUid]['id'],
@@ -5531,7 +5543,9 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
         extra_expr_ctx = _base_extra_expr_ctx(self)
         cfg = extra_expr_ctx['cfg']
         for org_uid, expr in cfg.get_orgs_with_as_copy_group_on_expression().items():
-            extra_expr_ctx.update({'item': self, 'isCreated': isCreated})
+            extra_expr_ctx.update({'item': self,
+                                   'isCreated': isCreated,
+                                   'org_uid': org_uid})
             suffixes = _evaluateExpression(
                 self,
                 expression=expr,
@@ -6077,7 +6091,7 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
     def mandatoryAdvicesAreOk(self):
         '''Returns True if all mandatory advices for this item have been given and are all positive.'''
         if not hasattr(self, 'isRecurringItem'):
-            for advice in self.adviceIndex.itervalues():
+            for advice in self.adviceIndex.values():
                 if not advice['optional'] and \
                    not advice['type'].startswith('positive'):
                     return False
@@ -6354,6 +6368,8 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
             org_uid for org_uid in self.adviceIndex
             if self.adviceIndex[org_uid].get('inherited', False) and
             org_uid not in handledAdviserUids]
+        # remove duplicates
+        unhandledAdviserUids = list(set(unhandledAdviserUids))
         if unhandledAdviserUids:
             optionalAdvisers += self.getUnhandledInheritedAdvisersData(
                 unhandledAdviserUids, optional=True)